pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vehron"
version = "0.1.0"
description = "VEHicle Research and Optimisation Network"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
  "numpy>=1.24",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "click>=8.1",
]
authors = [
  { name = "VEHRON Contributors" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]

[project.scripts]
vehron = "vehron.runner:cli"

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["vehron*"]
